chore: 파일 위치 깔끔하게 변경
