+++ src/main/java/gift/domain/member/Member.java
@@ -1,4 +1,4 @@
-package gift.domain;
+package gift.domain.member;
 
 import lombok.Builder;
 import lombok.Getter;
