+++ src/main/java/gift/repository/ProductRepository.java
@@ -1,5 +1,6 @@
-package gift.domain;
+package gift.repository;
 
+import gift.domain.Product;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.dao.EmptyResultDataAccessException;
