+++ src/main/java/gift/request/LoginRequest.java
@@ -1,18 +1,13 @@
-package gift.controller.request;
+package gift.request;
 
-import gift.domain.Member;
-import gift.error.ErrorMessage;
+import gift.constant.ErrorMessage;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
 import lombok.Getter;
 import org.hibernate.validator.constraints.Length;
 
 @Getter
-public class RegisterRequest {
-
-    @NotBlank(message = ErrorMessage.MEMBER_NAME_NOT_BLANK)
-    @Length(max = 15, message = ErrorMessage.MEMBER_NAME_EXCEEDS_MAX_LENGTH)
-    private String name;
+public class LoginRequest {
 
     @NotBlank(message = ErrorMessage.EMAIL_NOT_BLANK)
     @Pattern(regexp = "[a-zA-Z0-9_!#$%&’*+/=?`{|}~^.-]+@[a-zA-Z0-9.-]+$", message = ErrorMessage.INVALID_EMAIL_FORMAT)
@@ -21,13 +16,5 @@ public class RegisterRequest {
     @NotBlank(message = ErrorMessage.PASSWORD_NOT_BLANK)
     @Length(min = 4, max = 16, message = ErrorMessage.PASSWORD_LENGTH)
     private String password;
-
-    public Member toEntity() {
-        return Member.builder()
-            .name(this.name)
-            .email(this.email)
-            .password(this.password)
-            .build();
-    }
 
 }
