+++ src/main/java/gift/response/JwtResponse.java
@@ -1,4 +1,4 @@
-package gift.controller.response;
+package gift.response;
 
 import lombok.Getter;
