+++ src/main/java/gift/response/ProductResponse.java
@@ -1,4 +1,4 @@
-package gift.controller.response;
+package gift.response;
 
 public class ProductResponse {
