+++ src/main/java/gift/service/ProductService.java
@@ -1,11 +1,10 @@
 package gift.service;
 
-import gift.controller.request.ProductRequest;
-import gift.controller.response.ProductResponse;
+import gift.request.ProductRequest;
+import gift.response.ProductResponse;
 import gift.domain.Product;
-import gift.domain.ProductRepository;
-import gift.error.ErrorMessage;
-import gift.error.exception.ProductNotFoundException;
+import gift.repository.ProductRepository;
+import gift.exception.ProductNotFoundException;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
