+++ src/test/java/gift/domain/ProductRepositoryTest.java
@@ -2,7 +2,8 @@ package gift.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import gift.controller.request.ProductRequest;
+import gift.repository.ProductRepository;
+import gift.request.ProductRequest;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
